+++ src/character-sheet/characterContext.js
@@ -1,4 +1,4 @@
 import React from 'react';
-import { DEFAULT_CHARACTER } from './constants';
+import { DEFAULT_CHARACTER } from '../constants';
 
 export const characterContext = React.createContext(DEFAULT_CHARACTER);
