+++ src/character-sheet/left-column/AbilityScores.js
@@ -4,7 +4,7 @@ import {
   TextField,
 }
 from "@mui/material";
-import { Item } from './styles';
+import { Item } from '../../styles';
 
 function AbilityScores(props) {
