+++ src/character-sheet/left-column/LeftColumn.js
