+++ src/character-sheet/middle-column/CombatStats.js
@@ -2,7 +2,7 @@ import {
     Grid, 
     TextField 
 } from "@mui/material";
-import { Item } from "./styles";
+import { Item } from "../../styles";
 
 function CombatStats(props) {
