+++ src/character-sheet/middle-column/Equipment.js
