+++ src/character-sheet/middle-column/MiddleColumn.js
