+++ src/character-sheet/middle-column/Position.js
@@ -7,7 +7,7 @@ import {
     Typography 
 } from "@mui/material";
 import React from "react";
-import { Item } from "./styles";
+import { Item } from "../../styles";
 
 function Position(props) {
