+++ src/data/armors.json
