+++ src/data/weapons.json
