+++ src/home/Home.js
@@ -1,0 +1,8 @@
+function Home(props) {
+
+    return (
+      <>home</>
+    );
+  }
+  
+  export default Home;
