+++ src/login/Login.js
@@ -1,0 +1,8 @@
+function Login(props) {
+
+    return (
+      <>login</>
+    );
+  }
+  
+  export default Login;
