+++ src/register/Register.js
@@ -1,0 +1,8 @@
+function Register(props) {
+
+    return (
+      <>register</>
+    );
+  }
+  
+  export default Register;
