+++ src/rules/Rules.js
@@ -1,0 +1,8 @@
+function Rules(props) {
+
+    return (
+      <>rules</>
+    );
+  }
+  
+  export default Rules;
